requirements.txt
tqdm
scipy==1.7.3
opencv-python==4.10.0.82
mmcv==2.2.0
hydra-core==1.3.2
sagemaker==2.183.0